update: remote-device-job updator
