+++ projects/console-web-server/src/module/remote/remote-webdriver/remote-webdriver.controller.ts
@@ -16,30 +16,29 @@ import { Request, Response } from 'express';
 import { API_TOKEN_TYPE } from '../../auth/auth.types';
 import { ApiTokenPermission } from '../../auth/decorators';
 import { DoguLogger } from '../../logger/logger';
-import { WebDriverException } from '../../webdriver/webdriver.exception';
-import { WebDriverEndpointHandlerResult, WebDriverService } from '../../webdriver/webdriver.service';
+import { RemoteException } from '../common/exception';
+import { WebDriverEndpointHandlerResult } from '../common/type';
+import { RemoteWebDriverService } from './remote-webdriver.service';
 
 @Controller('/remote/wd/hub')
 export class RemoteWebDriverInfoController {
-  constructor(private readonly webdriverService: WebDriverService, private readonly logger: DoguLogger) {}
+  constructor(private readonly webdriverService: RemoteWebDriverService, private readonly logger: DoguLogger) {}
 
   @Post('session')
   @ApiTokenPermission(API_TOKEN_TYPE.WEBDRIVER_AGENT)
   async newSession(@Req() request: Request, @Res() response: Response): Promise<void> {
-    this.logger.debug(`@@@ newSession`);
     const relayRequest = this.webdriverService.convertRequest(request);
 
-    // parse capabilities then create endpoint and doguOptions
     const doguWebDriverCapabilitiesParser = new DoguWebDriverCapabilitiesParser();
     const endpoint = await WebDriverEndPoint.create(relayRequest, doguWebDriverCapabilitiesParser).catch((e) => {
-      throw new WebDriverException(400, e, {});
+      throw new RemoteException(400, e, {});
     });
     if (!doguWebDriverCapabilitiesParser.doguOptions) {
-      throw new WebDriverException(500, new Error('Internal error. doguOptions is null'), {});
+      throw new RemoteException(500, new Error('Internal error. doguOptions is null'), {});
     }
     const { doguOptions } = doguWebDriverCapabilitiesParser;
     if (endpoint.info.type !== 'new-session') {
-      throw new WebDriverException(400, new Error('Internal error. endpoint type is not new-session'), {});
+      throw new RemoteException(400, new Error('Internal error. endpoint type is not new-session'), {});
     }
 
     const processResult = await this.webdriverService.handleNewSessionRequest(endpoint.info, relayRequest, doguOptions);
@@ -49,17 +48,6 @@ export class RemoteWebDriverInfoController {
     this.setHeaders(headers, processResult);
     headers[DoguRequestTimeoutHeader] = DefaultHttpOptions.request.timeout3minutes.toString();
 
-    // FIXME: henry - temporary test
-    // headers[DoguRemoteDeviceJobIdHeader] = processResult.remoteDeviceJobId;
-    // headers[DoguRemoteDeviceJobIdHeader] = 'dogu';
-
-    // headers[DoguDevicePlatformHeader] = processResult.devicePlatform;
-    // headers[DoguDeviceSerialHeader] = processResult.deviceSerial;
-    // if (processResult.applicationUrl) headers[DoguApplicationUrlHeader] = processResult.applicationUrl;
-    // if (processResult.applicationVersion) headers[DoguApplicationVersionHeader] = processResult.applicationVersion;
-    // if (processResult.browserName) headers[DoguBrowserNameHeader] = processResult.browserName;
-    // if (processResult.browserVersion) headers[DoguBrowserVersionHeader] = processResult.browserVersion;
-
     const relayResponse = await this.webdriverService.sendRequest(processResult, headers);
     await this.webdriverService.handleNewSessionResponse(processResult, relayResponse);
     this.sendResponse(relayResponse, response);
@@ -67,13 +55,12 @@ export class RemoteWebDriverInfoController {
 
   @Delete('session/:sessionId')
   async deleteSession(@Req() request: Request, @Res() response: Response): Promise<void> {
-    this.logger.debug(`@@@ deleteSession`);
     const relayRequest = this.webdriverService.convertRequest(request);
     const endpoint = await WebDriverEndPoint.create(relayRequest).catch((e) => {
-      throw new WebDriverException(400, e, {});
+      throw new RemoteException(400, e, {});
     });
     if (endpoint.info.type !== 'delete-session') {
-      throw new WebDriverException(400, new Error('Internal error. endpoint type is not delete-session'), {});
+      throw new RemoteException(400, new Error('Internal error. endpoint type is not delete-session'), {});
     }
     const processResult = await this.webdriverService.handleDeleteSessionRequest(endpoint.info, relayRequest);
     const headers: HeaderRecord = {};
@@ -88,13 +75,12 @@ export class RemoteWebDriverInfoController {
   @All('session/:sessionId/*')
   // @ApiTokenPermission(API_TOKEN_TYPE.WEBDRIVER_AGENT)
   async process(@Req() request: Request, @Res() response: Response): Promise<void> {
-    this.logger.debug(`@@@ process`);
     const relayRequest = this.webdriverService.convertRequest(request);
     const endpoint = await WebDriverEndPoint.create(relayRequest).catch((e) => {
-      throw new WebDriverException(400, e, {});
+      throw new RemoteException(400, e, {});
     });
     if (endpoint.info.type !== 'session') {
-      throw new WebDriverException(400, new Error('Internal error. endpoint type is not session'), {});
+      throw new RemoteException(400, new Error('Internal error. endpoint type is not session'), {});
     }
     const processResult = await this.webdriverService.handleEachSessionRequest(endpoint.info, relayRequest);
     const headers: HeaderRecord = {};
